pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag_prep_tool"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "tool to prepare for rag."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


dependencies = [
  "PyMuPDF >= 1.24.4",
  "fast_antx @ git+https://github.com/OpenPecha/fast-antx.git",
  "llama-index-readers-file >= 0.1.23",
  "llama-index-core >= 0.10.40",
  "nltk >= 3.8.1",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
